Add files via upload
Project: Training and Clothing Marketing Platform

This university project focuses on the frontend development of a platform designed for training enthusiasts and clothing marketing. The application offers a visually engaging interface with features such as:

Simulated product browsing and selection.
Mock transaction flows to enhance the user experience.
Interactive and responsive design tailored for both desktop and mobile devices.
The project emphasizes the visual and user experience aspects of a modern e-commerce platform. Since it was developed for academic purposes, it does not include backend functionality or database integration.

This project showcases my skills in frontend development, UI/UX design, and the ability to simulate real-world applications effectively.